+++ ta4j-core/src/main/java/org/ta4j/core/BacktestExecutor.java
@@ -26,8 +26,8 @@ package org.ta4j.core;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.ta4j.core.cost.CostModel;
-import org.ta4j.core.cost.ZeroCostModel;
+import org.ta4j.core.analysis.cost.CostModel;
+import org.ta4j.core.analysis.cost.ZeroCostModel;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.reports.TradingStatement;
 import org.ta4j.core.reports.TradingStatementGenerator;