+++ ta4j-core/src/main/java/org/ta4j/core/Trade.java
@@ -26,8 +26,8 @@ package org.ta4j.core;
 import java.io.Serializable;
 import java.util.Objects;
 
-import org.ta4j.core.cost.CostModel;
-import org.ta4j.core.cost.ZeroCostModel;
+import org.ta4j.core.analysis.cost.CostModel;
+import org.ta4j.core.analysis.cost.ZeroCostModel;
 import org.ta4j.core.num.Num;
 
 /**